style(home-page): [WIP] - bugfix to searchmodal for accessibility #91
