+++ MusicBased_IOT_Platform/MusicBased_IOT_Platform/MusicBased_IOT_Platform/Components/Pages/Dashboard.razor
@@ -5,8 +5,8 @@
  @inject UserSessionService Session
 @inject NavigationManager NavigationManager
 
-<script src="https://d3js.org/d3.v7.min.js"></script>
-<script src="/js/dashboard.js"></script>
+@* <script src="https://d3js.org/d3.v7.min.js"></script>
+<script src="/js/dashboard.js"></script> *@
 
 <h3>My Dashboard (Biometric Trends)</h3>
 
@@ -19,6 +19,7 @@
     <option value="14">Weekly</option>
     <option value="30">Monthly</option>
 </select>
+
 
 
 @* d3 barchart *@
