+++ MusicBased_IOT_Platform/MusicBased_IOT_Platform/MusicBased_IOT_Platform/Components/Pages/Login.razor
@@ -52,7 +52,7 @@
                 </div>
             }
                     <div class="d-grid gap-2">
-                        <button class="btn btn-primary btn-lg" type="submit">Login</button>
+                        <button class="btn fitbit-aqua btn-lg" type="submit">Login</button>
                     </div>
             </div>
             </EditForm>
