+++ MusicBased_IOT_Platform/MusicBased_IOT_Platform/MusicBased_IOT_Platform/wwwroot/app.css
@@ -187,13 +187,30 @@ body {
 
 .search-container{
     position: relative;
-}
+/*    width: 100%;
+*/}
+
 
 .suggestions {
     position: absolute;
     top: 100%;
     left: 0;
+    right: 0;
+    margin-top: 5px;
+    background: white;
+    border-radius: 12px;
+    box-shadow: 0 8px 16px rgba(0,0,0,0.15);
+    max-height: 220px;
+    overflow-y: auto;
+    z-index: 2000;
+}
+
+/*.suggestions {
+    position: absolute;
+    top: 100%;
+    left: 0;
     width: 100%;
+
     background: white;
     border: 1px solid #e9ecef;
     border-radius: 10px;
@@ -201,16 +218,27 @@ body {
     z-index: 1050;
     max-height: 250px;
     overflow-y: auto;
-}
+}*/
 
 .suggestion-item {
-    padding: 10px 12px;
+    /*padding: 10px 12px;*/
+    padding: 12px 16px;
+    font-size: 15px;
     cursor: pointer;
 }
 
+    /*.suggestion-item:hover {
+        background: #f1f3f5;*/
+        /*#f5f5f5;*/
+    /*}*/
     .suggestion-item:hover {
-        background: #f5f5f5;
+        background: rgba(0,176,185,0.1);
     }
+
+.modal-body-scroll {
+    max-height: 70vh;
+    overflow-y: auto;
+}
 
 /*#suggestions {*/
 /*    position: absolute;
@@ -281,6 +309,25 @@ body {
   flex-direction: column;
   height: 100%;
 }*/
+
+.modal-content {
+    background: var(--app-white);
+    border-radius: 16px;
+}
+
+/*.modal-header {
+    background: var(--spotify-dark);
+    color: white;
+}*/
+
+.form-control {
+    border: 1px solid var(--app-border);
+}
+
+    .form-control:focus {
+        border-color: var(--spotify-green);
+        box-shadow: 0 0 0 0.2rem rgba(29,185,84,0.25);
+    }
 
 .card-body {
   flex-grow: 1;
@@ -367,6 +414,29 @@ body {
   transform: scale(1.1);
   box-shadow: 0px 4px 15px rgba(0, 0, 0, 0.2);
 }
+
+.go-btn {
+    background: var(--spotify-green);
+    color: white;
+    border: none;
+    width: 60px;
+    height: 60px;
+    font-size: 16px;
+    font-weight: bold;
+    border-radius: 0 12px 12px 0;
+    padding: 0 20px;
+    transition: transform 0.15s ease, box-shadow 0.15s ease;
+}
+    .go-btn:hover {
+        background: #17a74a;
+        transform: scale(1.05);
+        box-shadow: 0 4px 12px rgba(0,0,0,0.2);
+    }
+
+    .go-btn:active {
+        transform: scale(0.95);
+    }
+
 /*
 function getWindowScroll(node) {
     var win = getWindow(node);
